+++ packages/seroval/src/core/base/stream.ts
@@ -10,7 +10,7 @@ import type { BaseSyncParserContextOptions } from './sync';
 import BaseSyncParserContext from './sync';
 import { BIGINT_FLAG, Feature } from '../compat';
 import { SerovalNodeType, UNIVERSAL_SENTINEL } from '../constants';
-import { createRequestOptions, createResponseOptions } from '../constructors';
+import { createRequestOptions, createResponseOptions } from '../utils/constructors';
 import { NULL_NODE } from '../literals';
 import { serializeString } from '../string';
 import {
@@ -25,7 +25,7 @@ import {
   type SerovalResponseNode,
 } from '../types';
 import { createDOMExceptionNode, createURLNode, createURLSearchParamsNode } from '../web-api';
-import { asyncIteratorToReadableStream, iteratorToSequence } from '../iterator-to-sequence';
+import { asyncIteratorToReadableStream, iteratorToSequence } from '../utils/iterator-to-sequence';
 
 export interface BaseStreamParserContextOptions extends BaseSyncParserContextOptions {
   onParse: (node: SerovalNode, initial: boolean) => void;
