+++ packages/seroval/src/core/utils/iterator-to-sequence.ts
