+++ packages/seroval/src/core/utils/promise-to-result.ts
